backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.3
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
pytest==8.0.2
httpx==0.27.0
email-validator==2.1.0
psycopg2-binary 